src_website_quality/requirements.txt
torch
sentencepiece
tqdm
transformers==4.30.2
datasets
pillow
